versão de controle de histórico pág estática
