+++ resources/views/pages/home-manager.blade.php
@@ -19,11 +19,5 @@
             </div>
         </div>
     </div>
-    <div class="container">
-        <div class="row">
-            <div class="mt-lg-n12 mt-md-n13 mt-n12 justify-content-center">
-                {{-- @livewire('faq.faq-manager') --}}
-            </div>
-        </div>
-    </div>
+    
 @endsection
